// app/lib/types.ts
export type Artist = {
  //name?: string | undefined; // Fixed syntax error here
  name:string;
  picture?: string;
  info?: string;
  id: number|string | null;
  image?: string;
  gender?: string;
  followers?: number;
  monthlyListeners?: number;
  topSong?: string;
  genre?: string| string[];
}




// Main Track interface based on usage in the code
export type Track = {
  id: number;
  track_name: string;
  music_id?:string;
  genre: string;
  duration: string;
  label: string;
  image: string;
  actor?:string;
  artist?:Artist|undefined;
  plays: number;
 url?: string;
  //releaseDate?: date;
  releaseDate?: string;
  isLiked?: boolean;
  audio?: string;
  no_plays?:number;
  description?: string;
  info?: string;
}



// Define the player state type

export type PlayerStateType= {

  currentTrack: Track | null;
  isPlaying: boolean;
  isMuted?: boolean;
  isShuffle: boolean;
  repeatMode?: 'off' | 'one' | 'all';
  playbackSpeed: number;
  volume: number;
  queue: Track[];
  history: Track[];
  isFullScreen:boolean;
  currentIndex: number;
  currentTime: number;
  loading: boolean;
  error: string | null;
  playlist:Track[];
   currentTrackIndex: number;
  loopMode?: 'off' | 'one' | 'all';
}



export type PlayerActionType =
  | { type: 'SET_TRACK'; payload: Track }
  | { type: 'TOGGLE_PLAY' }
  | { type: 'TOGGLE_SHUFFLE' }
  | { type: 'PLAY' }
  | { type: 'SET_VOLUME'; payload: number }
  | { type: 'SET_REPEAT_MODE'; payload: 'off' | 'one' | 'all' }
  | { type: 'SET_PLAYBACK_SPEED'; payload: 1 | 1.5 | 2 }
  | { type: 'ADD_TO_QUEUE'; payload: Track }
  | { type: 'REMOVE_FROM_QUEUE'; payload: number }
  | { type: 'NEXT_TRACK' }
  | { type: 'PREVIOUS_TRACK' }
  | { type: 'SEEK_TO'; payload: number }
  | { type: 'LOAD_PLAYLIST'; payload: Track[] }
  | { type: 'TOGGLE_FULLSCREEN', payload?: boolean }
  | { type: 'SET_LOADING'; payload: boolean }
  | { type: 'SET_ERROR'; payload: string | null }
  | { type: 'UPDATE_TIME'; payload: number };